Test scenario

+++ src/producerConsumer/ClientParameters.java
@@ -2,12 +2,12 @@ package producerConsumer;
 
 public final class ClientParameters {
     public final int requests;
-    public final int calculations;
+    public final long calculations;
     public final int minPortion;
     public final int maxPortion;
     public final int operationsQuantum;
 
-    public ClientParameters(int requests, int calculations, int minConsumption, int maxConsumption, int operationsQuantum) {
+    public ClientParameters(int requests, long calculations, int minConsumption, int maxConsumption, int operationsQuantum) {
         this.requests = requests;
         this.calculations = calculations;
         this.minPortion = minConsumption;
